# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /platform/settings:
    get:
      operationId: getPlatformSettings
      tags: [platform]
      summary: Get owner trading settings
      responses:
        "200":
          description: Trading settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    patch:
      operationId: updatePlatformSettings
      tags: [platform]
      summary: Update safe trading settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
        "400":
          description: Invalid settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /platform/kill-switch:
    patch:
      operationId: updateKillSwitch
      tags: [risk]
      summary: Enable or disable the persistent kill switch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KillSwitchUpdate"
      responses:
        "200":
          description: Kill switch status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KillSwitchStatus"
...
  /portfolio:
    get:
      operationId: getPortfolio
      tags: [portfolio]
      summary: Get portfolio
      responses:
        "200":
          description: Portfolio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
    patch:
      operationId: updatePortfolio
      tags: [portfolio]
      summary: Update portfolio settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioBody"
      responses:
        "200":
          description: Updated portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: List open trades
      responses:
        "200":
          description: List of open trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: addTrade
      tags: [trades]
      summary: Analyze and add a trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "200":
          description: Trade analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /trades/{ticker}:
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Remove trade from tracking
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Trade not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /trades/{ticker}/status:
    patch:
      operationId: updateTradeStatus
      tags: [trades]
      summary: Pause or resume monitoring for a tracked asset
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, paused]
              required: [status]
      responses:
        "200":
          description: Trade status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Trade not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /symbols/search:
    get:
      operationId: searchSymbols
      tags: [analysis]
      summary: Search for a ticker by company name, ticker, ISIN, crypto or ETF name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching symbols
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SymbolSearchResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /analysis/{ticker}:
    get:
      operationId: analyzeTicker
      tags: [analysis]
      summary: Analyze a ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request or data fetch error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /signals/{id}/approve:
    post:
      operationId: approveSignal
      tags: [orders]
      summary: Manually approve a qualified LONG signal for execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approval result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalApproval"
...
components:
  schemas:
...
    PlatformSettings:
      type: object
      properties:
        riskPerTradePct:
          type: number
        maxExposurePct:
          type: number
        maxOpenPositions:
          type: integer
        maxDailyTrades:
          type: integer
        maxDailyLossPct:
          type: number
        cooldownMinutes:
          type: integer
        minRewardRisk:
          type: number
        paperStartingBalance:
          type: number
      required: [riskPerTradePct, maxExposurePct, maxOpenPositions, maxDailyTrades, maxDailyLossPct, cooldownMinutes, minRewardRisk, paperStartingBalance]
    PlatformSettingsUpdate:
      type: object
      properties:
        riskPerTradePct:
          type: number
          minimum: 0.01
          maximum: 1
        maxExposurePct:
          type: number
          minimum: 0.01
          maximum: 100
        maxOpenPositions:
          type: integer
          minimum: 1
          maximum: 2
        maxDailyTrades:
          type: integer
          minimum: 1
          maximum: 3
        maxDailyLossPct:
          type: number
          minimum: 0.01
          maximum: 2
        cooldownMinutes:
          type: integer
          minimum: 0
          maximum: 10080
        minRewardRisk:
          type: number
          minimum: 1.5
          maximum: 20
        paperStartingBalance:
          type: number
          minimum: 0
    KillSwitchUpdate:
      type: object
      properties:
        active:
          type: boolean
        reason:
          type: ["string", "null"]
      required: [active]
    KillSwitchStatus:
      type: object
      properties:
        active:
          type: boolean
        reason:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [active, updatedAt]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        orderType:
          type: string
          enum: [MARKET, LIMIT, STOP_LOSS_LIMIT, TAKE_PROFIT_LIMIT]
        status:
          type: string
          enum: [CREATED, SUBMITTED, ACKNOWLEDGED, PARTIALLY_FILLED, FILLED, CANCELED, REJECTED, UNKNOWN]
        clientOrderId:
          type: string
        quantity:
          type: number
        price:
          type: ["number", "null"]
        createdAt:
          type: string
      required: [id, symbol, side, orderType, status, clientOrderId, quantity, createdAt]
...
    SignalApproval:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        order:
          type: ["object", "null"]
          $ref: "#/components/schemas/Order"
      required: [success, message]
...
    Trade:
      type: object
      properties:
        ticker:
          type: string
        entry:
          type: number
        tp:
          type: number
        sl:
          type: number
        atr:
          type: number
        direction:
          type: string
          enum: [LONG]
        reason:
          type: string
        investAmount:
          type: number
        addedAt:
          type: string
        status:
          type: string
          enum: [active, paused]
      required: [ticker, entry, tp, sl, atr, direction, reason, investAmount, addedAt]
    ClosedTrade:
      type: object
      properties:
        ticker:
          type: string
        entry:
          type: number
        tp:
          type: number
        sl:
          type: number
        atr:
          type: number
        direction:
          type: string
          enum: [LONG]
        reason:
          type: string
        investAmount:
          type: number
        addedAt:
          type: string
        closedAt:
          type: string
        closeReason:
          type: string
          enum: [TP_HIT, SL_HIT, MANUAL]
        exitPrice:
          type: number
        pnl:
          type: number
      required: [ticker, entry, tp, sl, atr, direction, reason, investAmount, addedAt, closedAt, closeReason, exitPrice, pnl]
    TradeInput:
      type: object
      properties:
        ticker:
          type: string
        investAmount:
          type: number
      required: [ticker]
    SymbolSearchMatch:
      type: object
      properties:
        symbol:
          type: string
        instrumentName:
          type: string
        exchange:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        instrumentType:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
      required: [symbol, instrumentName]
    SymbolSearchResult:
      type: object
      properties:
        query:
          type: string
        matches:
          type: array
          items:
            $ref: "#/components/schemas/SymbolSearchMatch"
        providerLimited:
          type: boolean
        note:
          type: ["string", "null"]
      required: [query, matches, providerLimited]
    TradeResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        signal:
          type: string
        direction:
          type: ["string", "null"]
        score:
          type: ["number", "null"]
        verdict:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        heuristicConfidence:
          type: ["number", "null"]
        reason:
          type: ["string", "null"]
        trade:
          $ref: "#/components/schemas/Trade"
      required: [success, message]
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    Portfolio:
      type: object
      properties:
        balance:
          type: number
        telegramChatId:
          type: ["string", "null"]
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        closedTrades:
          type: array
          items:
            $ref: "#/components/schemas/ClosedTrade"
      required: [balance, trades, closedTrades]
    UpdatePortfolioBody:
      type: object
      properties:
        balance:
          type: number
        telegramChatId:
          type: string
...
    ScoreBreakdown:
      type: object
      properties:
        trend:
          type: number
        momentum:
          type: number
        volatility:
          type: number
        volume:
          type: number
        structure:
          type: number
        multiTimeframe:
          type: number
      required: [trend, momentum, volatility, volume, structure, multiTimeframe]
    MtfAnalysis:
      type: object
      properties:
        m15:
          type: string
        h1:
          type: string
        h4:
          type: string
        daily:
          type: string
      required: [m15, h1, h4, daily]
    AnalysisResult:
      type: object
      properties:
        ticker:
          type: string
        price:
          type: number
        score:
          type: number
        signal:
          type: string
        direction:
          type: string
          enum: [LONG, WAIT]
        verdict:
          type: string
          enum: [FORTE_BUY, BUY, NEUTRALE]
        reason:
          type: string
        tp:
          type: number
        sl:
          type: number
        atr:
          type: number
        ema50:
          type: number
        ema100:
          type: number
        ema200:
          type: number
        rsi:
          type: number
        macdHistogram:
          type: number
        volumeRatio:
          type: number
        falseSignalRisk:
          type: string
          enum: [Basso, Medio, Alto]
        confidenceScore:
          type: number
        heuristicConfidence:
          type: number
        scoreBreakdown:
          $ref: "#/components/schemas/ScoreBreakdown"
        mtfAnalysis:
          $ref: "#/components/schemas/MtfAnalysis"
        invalidationConditions:
          type: array
          items:
            type: string
      required: [ticker, price, score, signal, direction, verdict, reason, tp, sl, atr, ema50, ema100, ema200, rsi, macdHistogram, volumeRatio, falseSignalRisk, confidenceScore, heuristicConfidence, scoreBreakdown, mtfAnalysis, invalidationConditions]
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]